src/python/data-ingestion/pyproject.toml
[tool.poetry]
name = "data-ingestion"
version = "0.1.0"
description = "Pipeline to ingest data from scrappers to RDS"
authors = ["Erik Dao <cuong@neuraltalks.io>"]
packages = [{include = "data_ingestion", from = "src"}]
classifiers = [
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.9",
]

[tool.poetry.dependencies]
python = "^3.9"
polars = "^0.17.5"
pyarrow = "^11.0.0"
sqlalchemy = "^2.0.9"
pandas = "^2.0.0"
psycopg2-binary = "^2.9.6"

[tool.semantic_release]
branch = "main"
version_variable = "src/data_ingestion/__init__.py:__version__"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
